refactor: 컴포넌트 폴더 리팩토링
