+++ src/Components/common/ImageDetailView.tsx
