+++ src/Pages/MainPage.tsx
@@ -1,5 +1,5 @@
 import NavBar from "../Components/common/NavBar";
-import Intro from "../Components/Intro";
+import Intro from "../Components/introduce/Intro";
 import Footer from "../Components/common/Footer";
 import FAQ from "../Components/common/FAQ";
 const MainPage = () => {
