+++ src/Pages/NoticeDetailPage.tsx
@@ -5,7 +5,7 @@ import { useEffect, useState } from "react";
 import Notice from "../Types/NoticeType";
 import { doc, getDoc } from "firebase/firestore";
 import { db } from "../firebase";
-import LoadingCircle from "../Components/LoadingCircle";
+import LoadingCircle from "../Components/common/LoadingCircle";
 import { FaLongArrowAltLeft } from "react-icons/fa";
 
 const NoticeDetailPage = () => {
