+++ src/Pages/ProjectDetailPage.tsx
@@ -24,11 +24,11 @@ import DOMPurify from "dompurify";
 import { User, getAuth } from "firebase/auth";
 import CheckDialog from "../Components/CheckDialog";
 import ProjectType, { ProjectDetail } from "../Types/ProjectType";
-import Comment from "../Components/Comment";
+import Comment from "../Components/project/Comment";
 import CommentType from "../Types/CommentType";
 import { HiHeart, HiShare, HiOutlineHeart } from "react-icons/hi2";
-import LoadingCircle from "../Components/LoadingCircle";
-import ImageDetailView from "../Components/ImageDetailView";
+import LoadingCircle from "../Components/common/LoadingCircle";
+import ImageDetailView from "../Components/common/ImageDetailView";
 import ScrollToTop from "../Components/common/ScrollToTop";
 import { IoLogoGithub } from "react-icons/io";
 import { HiLink } from "react-icons/hi";
